Split openai provider from openai-compatible
The single openai.ts conflated cloud OpenAI (api.openai.com curated
catalog) with any-3rd-party-OpenAI-API server (Ollama, LM Studio, vLLM,
llama.cpp). They have different concerns: cloud OpenAI has known
reasoning quirks per model family (o-series, GPT-5, GPT-5.1+), while
openai-compatible has no canonical reasoning shape because it depends
on what model the local server is serving.

Split into two extensions, each gated on its own env var:
- openai            → OPENAI_API_KEY without OPENAI_BASE_URL
- openai-compatible → OPENAI_BASE_URL

Setting both still routes to openai-compatible, preserving prior
behavior. Manifest names give users finer disabledBuiltins control.
