+++ src/extensions/index.ts
@@ -26,8 +26,11 @@ export const BUILTIN_EXTENSIONS: Array<{
     when: () => !!process.env.OPENROUTER_API_KEY,
     load: () => import("./providers/openrouter.js").then(m => m.default) },
   { name: "openai",
-    when: () => !!process.env.OPENAI_API_KEY,
+    when: () => !!process.env.OPENAI_API_KEY && !process.env.OPENAI_BASE_URL,
     load: () => import("./providers/openai.js").then(m => m.default) },
+  { name: "openai-compatible",
+    when: () => !!process.env.OPENAI_BASE_URL,
+    load: () => import("./providers/openai-compatible.js").then(m => m.default) },
   { name: "deepseek",
     load: () => import("./providers/deepseek.js").then(m => m.default) },
   { name: "tui-renderer",     load: () => import("./tui-renderer.js").then(m => m.default) },
